dom/base/test/meta_viewport/test_meta_viewport_auto_size_by_initial_scale_1.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>initial-scale=1 without width enables autoSize</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="initial-scale=1">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>initial-scale=1</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function initial_scale_1() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      ok(info.autoSize,
         "initial-scale=1 without width should enable autoSize");
    });
  </script>
</body>
</html>
